pyproject.toml
[tool.poetry]
name = "linkedinscrapper"
version = "0.1.0"
description = ""
authors = ["Your Name <you@example.com>"]
readme = "README.md"

[tool.poetry.dependencies]
python = "^3.12"
selenium = "^4.26.1"
psycopg2 = "^2.9.10"
redis = "^5.2.0"
python-dotenv = "^1.0.1"
pandas = "^2.2.3"
matplotlib = "^3.9.2"
seaborn = "^0.13.2"


[build-system]
requires = ["poetry-core"]
build-backend = "poetry.core.masonry.api"
